Fixed after code review
